backend/requirements.txt
Django==2.2.19
pytz==2022.7.1
sqlparse==0.4.3
django-colorfield==0.8.0
djangorestframework==3.9.2
drf-extra-fields==3.4.1
gunicorn==20.0.4
psycopg2-binary==2.8.6
python-dotenv==0.21.1
asgiref==3.2.10
django-filter==2.4.0
djangorestframework-simplejwt==4.8.0
pyJWT==2.1.0
djoser==2.1.0
django-cors-headers==2.5.0